test #3

+++ config/database.php
@@ -81,8 +81,9 @@ return [
 
         'pgsql' => [
             'driver' => 'pgsql',
+            // Explicitly disable DATABASE_URL parsing by not including 'url' key
             'host' => env('DB_HOST', env('PGHOST', '127.0.0.1')),
-            'port' => env('DB_PORT', env('PGPORT', '5432')),
+            'port' => (int) env('DB_PORT', env('PGPORT', '5432')),
             'database' => env('DB_DATABASE', env('PGDATABASE', 'laravel')),
             'username' => env('DB_USERNAME', env('PGUSER', 'root')),
             'password' => env('DB_PASSWORD', env('PGPASSWORD', '')),
